Test failed: Use file-backed iSCSI targets insted of zvol, because zvols are not included in TruaNAS Backup
TrueNAS Cloud Backup basks up Data to a Cloud S3 Service. However the Backup is file based and zvols in ZFS exist as a childlren of Datasets but not actual files, therefore Cloud-Backup does not see them.

Using file-backed iSCSI targets instead of zvols works. But Cloud-Backup sees the full logical size and tried to Backup the whole terabyte-file instead of the physical file size.

The official way is to a export the zvol-snapshot out of ZFS to a file, this stores the zvol twice